[fix] some typos fixed. [ref] templates changed. dockerfiles changed

+++ setup.sh
@@ -1,18 +1,28 @@
-source config.example 
+source config
+if [ ! -f config ]; then
+    echo "Error: config file not found!"
+    echo "Use command 'cp config.example config' then proceed"
+    exit 1
+fi 
+echo "Loaded configuration from config file."
 
+echo "Running fetch-params.sh script..."
 mkdir -p /home/${USER}/.zcash-params
 ./fetch-params.sh
 
+echo "Generating .env file with MCL version and user details..."
 echo "MCL_VERSION=${mcl_version}" > .env
+echo "USER=${USER}" >> .env
 echo "USER_ID=$(id -u)" >> .env
 echo "GROUP_ID=$(id -g)" >> .env
+#echo ".env file created."
 
-OUTPUT_FILE="docker-compose_t.yml"
-VOLUME_TEMPLATE="templates/volume.template"
+OUTPUT_FILE="docker-compose.yml"
 SERVICE_TEMPLATE="templates/docker-service.template"
 
 # Function to check for duplicate values in an array
 check_duplicates() {
+    echo "Checking for duplicate node names and SSH ports..."
     declare -A seen_names
     declare -A seen_ports
     for node_info in "${nodes[@]}"; do
@@ -30,22 +40,41 @@ check_duplicates() {
         fi
         seen_ports[$ssh_port]=1  # Mark the SSH port as seen
     done
+    echo "No duplicates found. Proceeding..."
 }
-check_duplicates 
+check_duplicates
 
-cp "$VOLUME_TEMPLATE" "$OUTPUT_FILE"
-echo "sevices:" >> "$OUTPUT_FILE"
+if [ -f "$OUTPUT_FILE" ]; then
+    echo "Backing up existing docker-compose.yml to docker-compose.yml.bak"
+    mv "$OUTPUT_FILE" "${OUTPUT_FILE}.bak"
+fi
+
+echo "Starting to create docker compose file from template files ..."
+echo "version: '3.7'" >> "$OUTPUT_FILE"
+echo "services:" >> "$OUTPUT_FILE"
 
 for node_info in "${nodes[@]}"; do
     cat "$SERVICE_TEMPLATE" >> "$OUTPUT_FILE"
-    IFS=":" read -r node_name password ssh_port pubkey  <<< "$node_info"
+    IFS=":" read -r node_name ssh_user ssh_pw ssh_port pubkey  <<< "$node_info"
 
-    sed -i "s/{{ node.name }}/$node_name/g" "$OUTPUT_FILE"
-    sed -i "s/{{ container.name }}/$node_name/g" "$OUTPUT_FILE"
-    sed -i "s/{{ node.password }}/$password/g" "$OUTPUT_FILE"
-    sed -i "s/{{ ssh_port }}/$ssh_port/g" "$OUTPUT_FILE"
-    sed -i "s/{{ volume.name }}/$node_name/g" "$OUTPUT_FILE"
-    sed -i "s/{{ node.pubkey }}/$pubkey/g" "$OUTPUT_FILE"
+    ssh_user_var=$node_name"_ssh_user"
+    ssh_pw_var=$node_name"_ssh_pw"
+    echo "$ssh_user_var=$ssh_user" >> .env
+    echo "$ssh_pw_var=$ssh_pw" >> .env
 
+    sed -i -e "s/{{ node.name }}/$node_name/g" \
+           -e "s/{{ container.name }}/$node_name/g" \
+           -e "s/{{ ssh_user }}/\${$ssh_user_var}/g" \
+           -e "s/{{ ssh_pw }}/\${$ssh_pw_var}/g" \
+           -e "s/{{ ssh_port }}/$ssh_port/g" \
+           -e "s/{{ container.user }}/\${$ssh_user_var}/g" \
+           -e "s/{{ volume.name }}/$node_name/g" \
+           -e "s/{{ volume.user }}/\${$ssh_user_var}/g" \
+           -e "s/{{ node.pubkey }}/$pubkey/g" "$OUTPUT_FILE"
+
+
+    echo "Creating directory chainfiles/$node_name..."
     mkdir -p chainfiles/$node_name
 done
+
+echo "Docker Compose file generation complete. All steps finished successfully."
